+++ GreenShells/GreenShells/PlayerRemote.cpp
@@ -387,9 +387,9 @@ PlayerRemote* PlayerRemote::Deserialize(boost::property_tree::ptree playerNode)
     player->m_playerID = playerNode.get<int>("<xmlattr>.PId");
     player->m_playerName = playerNode.get<std::string>("<xmlattr>.PName");
     player->m_unitCount = playerNode.get<int>("<xmlattr>.UC");
-    player->m_food = playerNode.get<int>("<xmlattr>.F");
-    player->m_science = playerNode.get<int>("<xmlattr>.S");
-    player->m_weapon = playerNode.get<int>("<xmlattr>.W");
+    player->m_food = playerNode.get<double>("<xmlattr>.F");
+    player->m_science = playerNode.get<double>("<xmlattr>.S");
+    player->m_weapon = playerNode.get<double>("<xmlattr>.W");
     player->m_foodMultiplier = playerNode.get<double>("<xmlattr>.FM");
     player->m_scienceMultiplier = playerNode.get<double>("<xmlattr>.SM");
     player->m_weaponMultiplier = playerNode.get<double>("<xmlattr>.WM");
